Revert "(CR):[velar]merge GKI android14-6.1-2025-05_r2 for LTS upgrade"
This reverts commit 5e2a6a6d1f61dd60db12c09203681a58576bd6cb.

Reviewed-on: https://gerrit.mot.com/3613297
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Submit-Approved: Deise Alves <deisema@motorola.com>

+++ drivers/gpu/drm/amd/amdgpu/amdgpu_eeprom.h
@@ -28,10 +28,10 @@
 
 int amdgpu_eeprom_read(struct i2c_adapter *i2c_adap,
 		       u32 eeprom_addr, u8 *eeprom_buf,
-		       u32 bytes);
+		       u16 bytes);
 
 int amdgpu_eeprom_write(struct i2c_adapter *i2c_adap,
 			u32 eeprom_addr, u8 *eeprom_buf,
-			u32 bytes);
+			u16 bytes);
 
 #endif
